requirements.txt
rumps==0.4.0
pynput==1.7.6
python-dotenv==1.0.1
pyperclip==1.9.0
google-genai==0.2.2